Create http-fetcher extension.kt file
Aligning the structure with async-fetcher
